pyproject.toml
[project]
name = "imposter-syndrome-ndif"
version = "0.1.0"
description = "Imposter Syndrome: LoRA fine-tuning for AI deception red teaming (NDIF/Cadenza Labs)"
requires-python = ">=3.10"
dependencies = [
    "transformers>=4.45.0",
    "peft>=0.13.0",
    "trl>=0.12.0",
    "datasets>=3.0.0",
    "bitsandbytes>=0.44.0",
    "accelerate>=1.0.0",
    "pyyaml>=6.0",
    "nnsight>=0.3.0",
    "torch>=2.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"
